search/requirements.txt
python-dotenv
typesense
PyYAML
paho-mqtt
